chap-backend/pyproject.toml
[tool.poetry]
name = "chap-backend"
version = "0.1.0"
description = ""
authors = ["Arnav Komaragiri <komaraac@mail.uc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
transformers = "^4.38.1"
accelerate = "^0.27.2"
langchain = "^0.1.9"
lxml = "^5.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
